requirements.txt
torch==1.11.0
gym==0.19.0
numpy
prodict
pyyaml
matplotlib
pandas
scipy
tqdm
